Some better life

+++ src/main/java/org/greenblitz/motion/base/IEncoder.java
@@ -34,7 +34,7 @@ public interface IEncoder {
      *
      * @param ticks
      */
-    void setTicksPerMeter(int ticks);
+    void setTicksPerMeter(double ticks);
 
-    int getTicksPerMeter();
+    double getTicksPerMeter();
 }
